+++ crates/messaging/affinidi-messaging-mediator/src/messages/store.rs
@@ -62,6 +62,7 @@ async fn _store_message(
     state
         .monitor
         .stored(&msg_id, from_did_hash, to_did_hash, data, live);
+    state.activity.received(state, to_did_hash).await;
     Ok(msg_id)
 }
 
@@ -392,6 +393,7 @@ pub(crate) async fn store_forwarded_message(
                 state
                     .monitor
                     .stored(&msg_id, sender_hash, &recipient_did_hash, message, live);
+                state.activity.received(state, &recipient_did_hash).await;
                 debug!(
                     "message id({}) stored successfully recipient({})",
                     msg_id, recipient